+++ Models/IMhwItem.cs
@@ -3,7 +3,7 @@
 namespace MHW_Editor.Models {
     public interface IMhwItem : INotifyPropertyChanged {
         byte[] Bytes { get; }
-        int Offset { get; }
+        ulong Offset { get; }
         string Name { get; }
         bool Changed { get; }
     }
