+++ Weapons/BowGun.cs
@@ -2,7 +2,7 @@
 
 namespace MHW_Editor.Weapons {
     public class BowGun : Ranged {
-        public BowGun(byte[] bytes, int offset, string weaponFilename) : base(bytes, offset, weaponFilename) {
+        public BowGun(byte[] bytes, ulong offset, string weaponFilename) : base(bytes, offset, weaponFilename) {
         }
 
         public new SpecialAmmo Special_Ammo_Type {
